requirements.txt
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9
SQLAlchemy==1.4.18
Werkzeug==2.0.1
python-dateutil==2.8.2
pytz==2021.1